have clojure-ns accept class names without $
When using aleph in combination with ring's wrap-stacktrace, certain
exceptions are silently swallowed by the server.  I traced this down
to clojure-ns raising a null pointer exception for the class-name
"lamina.core.observable.ConstantObservable", which is a defrecord.

This changes clojure-ns to return "lamina.core.observable" as the ns.
clojure-fn is unchanged and returns the full string, which is probably
wrong.